Add mapkick

+++ db/migrate/20230304095228_create_coordinates.rb
@@ -1,8 +1,8 @@
 class CreateCoordinates < ActiveRecord::Migration[7.0]
   def change
     create_table :coordinates do |t|
-      t.integer :lat
-      t.integer :long
+      t.decimal :lat
+      t.decimal :long
       t.belongs_to :sighting, null: false, foreign_key: true
 
       t.timestamps
